[exec] rearranged
