+++ includes/commands.h
@@ -6,7 +6,7 @@
 /*   By: feralves <feralves@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/20 10:32:11 by joapedr2          #+#    #+#             */
-/*   Updated: 2023/05/22 14:21:05 by feralves         ###   ########.fr       */
+/*   Updated: 2023/05/22 16:50:55 by feralves         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ char	*cmd_path(char *command);
 void	free_cmd(void);
 
 // execute/commands_list.c
-int		run_command(void);
+int		run_line(void);
 
 //execute/commands_builtins.c
 int		execute_builtin(char **input, int check);
@@ -42,5 +42,8 @@ int		is_builtin(char *input);
 
 //execute/executables.c
 int		executables_files(void);
+
+//execute/commands_exec.c
+void	exeggutor(t_cmd *cmd);
 
 #endif //COMMANDS_H
