--- test2
@@ -1,12 +1,0 @@
-total 392$
--rw-r--r--  1 ferfahl ferfahl   4033 may 22 15:04 makefile$
-drwxr-xr-x  2 ferfahl ferfahl   4096 may 21 02:56 includes$
-drwxr-xr-x  3 ferfahl ferfahl   4096 Apr 19 18:12 libs$
--rwxr-xr-x  1 ferfahl ferfahl 363168 may 22 15:24 minishell$
-drwxr-xr-x 12 ferfahl ferfahl   4096 may 22 15:24 objects$
--rw-r--r--  1 ferfahl ferfahl    197 Apr 19 18:12 readline.supp$
-drwxr-xr-x 12 ferfahl ferfahl   4096 may 18 10:44 srcs$
--rw-r--r--  1 ferfahl ferfahl      3 may 22 15:24 test1$
--rw-r--r--  1 ferfahl ferfahl      0 may 22 15:24 test2$
--rw-r--r--  1 ferfahl ferfahl     12 may 22 15:02 test3$
--rw-r--r--  1 ferfahl ferfahl     66 may 22 15:00 teste$
