--- test3
@@ -1,1 +1,0 @@
-hi1 hi2 hi3
