--- teste
@@ -1,8 +1,0 @@
-Makefile
-includes
-libs
-minishell
-objects
-readline.supp
-srcs
-teste
